House_price_data
A cross-validated k-nearest neighbor model from the house price data, and produces the following:

A 3-D plot of the house price for average floor area. A 3-D plot of the house price for floor areas of 75 square metres. A 3-D plot of the house price for floor areas of 125 square meters.